backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
anthropic>=0.84,<1.0
